+++ src/dbally/nl_responder/prompts.py
@@ -75,7 +75,7 @@ class QueryExplanationPromptFormat(PromptFormat):
 
 
 NL_RESPONSE_TEMPLATE = PromptTemplate[NLResponsePromptFormat](
-    chat=(
+    [
         {
             "role": "system",
             "content": "You are a helpful assistant that helps answer the user's questions "
@@ -90,11 +90,11 @@ NL_RESPONSE_TEMPLATE = PromptTemplate[NLResponsePromptFormat](
             "content": "The table below represents the answer to a question: {question}.\n"
             "{results}\nAnswer the question: {question}.",
         },
-    )
+    ],
 )
 
 QUERY_EXPLANATION_TEMPLATE = PromptTemplate[QueryExplanationPromptFormat](
-    chat=(
+    [
         {
             "role": "system",
             "content": "You are a helpful assistant that helps describe a table generated by a query "
@@ -113,5 +113,5 @@ QUERY_EXPLANATION_TEMPLATE = PromptTemplate[QueryExplanationPromptFormat](
             "Describe the table generated using this query: {query}.\n"
             "Number of results to this query: {number_of_results}.\n",
         },
-    )
+    ],
 )
