+++ src/dbally/view_selection/llm_view_selector.py
@@ -5,7 +5,7 @@ from dbally.llms.base import LLM
 from dbally.llms.clients.base import LLMOptions
 from dbally.prompt.template import PromptTemplate
 from dbally.view_selection.base import ViewSelector
-from dbally.view_selection.view_selector_prompt_template import VIEW_SELECTION_TEMPLATE, ViewSelectionPromptFormat
+from dbally.view_selection.prompt import VIEW_SELECTION_TEMPLATE, ViewSelectionPromptFormat
 
 
 class LLMViewSelector(ViewSelector):
